Player Fix
Level is now a real abstract class
These methods are now pure virtual:
virtual Player* SetPlayer(Player* _player) = 0;
virtual Level* CreateLevel() = 0;
They have to be implemented in a derived class. There is also a help function now in Level called SetPlayerPosition. This method will set the spawn position of the player, it will also create a new player if it is not created yet and otherwise use the previous player and delete the weapons.