int score entity fix because of code review
entity does need SetScore and GetScore, was attempting to make it pure virtual. But in this case, that's not a good idea, both in implementing it and theoretically regarding infrastructure/setup of the classes. Programming -> brings problems. Rightfully so, because pure virtual means every class needs it and that is not true. Only NPC (and maybe Player)